Add files via upload
"img2audio_one_row" shows how the logic is implemented over only the first row of pixels in the picture. "img2audio".
In both the files, the image is converted to a 320x480 black and white picture before the process is initiated.
Once the image is obtained in the desired resolution and format, the pixel intensity values of each row are then converted into an audio signal with each pixel corresponding to a frequency, and the pixel intensity corresponds to the component magnitude. Each row of pixels is converted to sound signal and then stitched together, with a separator signal in between each row.